♻️ SidePanel 공용 컴포넌트로 분리
- 내부 콘텐츠는 children으로 받도록 처리
